Fixed Terrain & bioms placing while rotating
Bioms and Terrain no longer place while rotating the planet. In addition, terrain and bioms cannot be pained while 2 fingers are on the screen.

Added a TestingControls script to the camera, which includes the painting and rotation functionality with the arrow keys and mouse controls. This is for testing purposes and will not be included in the final build.