Atmosphere Controls Only While Atmosphere Selected
Created 2 methods in AtmosphereController.cs to be accessed by buttons. void ActivateAtmosphere() sets a bool to true, while void DeactivateAtmosphere() sets the bool to false.

Can only use atmosphere tools if bool is true. Buttons for Terrain and Biomes call the method DeactivateAtmosphere to change it to false. Buttons Atmosphere and any of the atmosphere pain buttons call ActivateAtmosphere to change it to true.