Terrain Assets Included
Imported current 3D terrain models and texutres. Created terrain object prefabs and prefab varients of each 3D terrain model. Included prefabs in painter and testing, but pine and hill are the only ones currently work. Pines and hills currently change to snowy versions on snow and grassy versions on grass no matter if the biome or the terrain was placed first. Will need to implement the other models in the future.