Compass Cube Rotates
Compass Cube added to compass in Gas Giant Creator scene. Rotates with planet, is set back to its original position and rotation upon clicking Rotation 0 Button.

Rotation 0 Button in Gas Giant Creator scene given its own Canvas and set to World Space to make this work. This is only in the Gas Giant Creator scene, so that the feature can be easily reverted to the settings of the Terrestrial Creator Scene if it is disliked.